Creates after_create hook when extraction is created to build out eefps's.
However we cannot use it yet as there is no reliable means of ensuring
database consistency if linkage between type1 and type2 is changed after
the after_create hook is called. We can make use of this once we have a
way to verify consistency...or fix it after the fact.