508 compliance fix:
Complaint:
The home page has 49 headings listed.  There's a lot of content that's marked as heading that should not be marked as a heading.
The H5 and H6 headings on the home page can be removed as headings.  Follow a similar structure to the one used for the "About"
page when creating the headings for the "Home" page.

Resolution:
Removed h5 and h6 headings and replaced with span. Also substitute
styles in spreadsheet.