508 compliance fix:
Complaint:
When in the edit profile section of the website if you hit enter on the edit profile button the keyboard
focus jumps back the very beginning of the page instead of into the fields to edit your profile.
This also appears to happen every time you try enter into any section within the purple header section.
It always jump you back to the beginning of the page.

Resolution:
Placed autofocus: true on first element as per HTML5.