Update citation_type.rb
CitationType should never be deleted. I think it is best that if it does we do not do anything to its descendants. This will raise fk exception and if that somehow fails too at least the children will retain their parent ID and it would be simpler to recover the parent and the children will be fine again.